First complete iteration of exercise 1
Base code of exercise 1.
Currently bugged:
Exception in thread "main" java.lang.NullPointerException
	at topic1.ItemList.addToItemList(ItemList.java:15)
	at topic1.ShoppingCart.setItemList(ShoppingCart.java:23)
	at topic1.Demo.main(Demo.java:11)

+++ Open training/src/topic1/ItemBuilder.java
@@ -4,7 +4,7 @@ public interface ItemBuilder {
 
 	Item  getItem();
 	
-	ItemBuilder buildPrice(int price);
+	ItemBuilder buildPrice(double price);
 	
 	ItemBuilder buildName(String name);
 }
